+++ CHANGELOG.md
@@ -85,6 +85,7 @@ The format is based on [Keep a Changelog 1.0.0].
 - add `skipped_ids` to `Task` to allow users to skip questions
 - add error summary component
 - customise presence validation message
+- provide qualtrics feedback link
 
 **Multiple Categories**